Refactoring & Childrouting fix for ad-detail-component
Refactoring & Childrouting fix for ad-detail-component
