+++ src/app/app.module.ts
@@ -12,7 +12,7 @@ import { NavigationMenuComponent } from './components/navigation-menu/navigation
 import { RegisterComponent } from './components/register/register.component';
 import { AdvertisementIndexComponent } from './components/advertisement/advertisement-index/advertisement-index.component';
 import { AdvertisementCreateComponent } from './components/advertisement/advertisement-create/advertisement-create.component';
-import { AdvertisementDetailComponent } from './components/advertisement-detail/advertisement-detail.component';
+import { AdvertisementDetailComponent } from './components/advertisement/advertisement-detail/advertisement-detail.component';
 import { AboutSummaryComponent } from './components/about/about-summary/about-summary.component';
 import { AboutUseCasesComponent } from './components/about/about-use-cases/about-use-cases.component';
 import { AdvertisementComponent } from './components/advertisement/advertisement.component';
