+++ src/app/components/advertisement/advertisement-detail/advertisement-detail.component.spec.ts
