+++ src/app/components/advertisement/advertisement-index/advertisement-index.component.html
@@ -1,11 +1,13 @@
   <!-- advertisement-index.component.html --> 
 
   <div class="container" style="width: 100%">
-    <div class="row" style="margin: 1pt; width: 150%">
-  <div style="margin: 1pt; border: 1px lightgray solid" class="col-12" *ngFor="let advertisement of (advertisements$ | async)">
-        <h3><a href="javascript:void(0)" routerLink="detail/{{ advertisement._id }}">{{advertisement.title}}</a></h3>
+    <div class="row" style="margin: 1pt; width: 100%">
+      <div style="margin: 1pt; border: 1px lightgray solid" class="col-12" *ngFor="let advertisement of (advertisements$ | async)">
+        <h3><a href="javascript:void(0)" [routerLink]="['/advertisement/detail', advertisement._id]">{{advertisement.title}}</a></h3>
         <p>{{advertisement.content.substring(0,150)}}</p>
-        <small><p>Author: {{advertisement.username}}</p></small>
-  </div>
+        <small>
+          <p>Author: {{advertisement.username}}</p>
+        </small>
+      </div>
     </div>
   </div>
