+++ src/app/components/advertisement/advertisement.component.html
@@ -2,11 +2,11 @@
 
 <div class="text-lg-center">
     <h1 class="display-4">Advertisement</h1>
+    <p class="lead">
+      <a [routerLink]="['./index']">Index</a> |
+      <a [routerLink]="['./create']">create</a>
+    </p>
     <hr class="my-4">
-</div>
-<p class="lead">
-    <a [routerLink]="['./index']">Index</a> |
-    <a [routerLink]="['./create']">create</a> 
-</p>
-
+  </div>
+  
 <router-outlet></router-outlet>
